+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/bankauthrepair/BankAuthRepairSubcomponent.kt
@@ -1,6 +1,6 @@
 package com.stripe.android.financialconnections.features.bankauthrepair
 
-import com.stripe.android.financialconnections.features.partnerauth.PartnerAuthState
+import com.stripe.android.financialconnections.features.partnerauth.SharedPartnerAuthState
 import dagger.BindsInstance
 import dagger.Subcomponent
 
@@ -13,7 +13,7 @@ internal interface BankAuthRepairSubcomponent {
     interface Builder {
 
         @BindsInstance
-        fun initialState(initialState: PartnerAuthState): Builder
+        fun initialState(initialState: SharedPartnerAuthState): Builder
 
         fun build(): BankAuthRepairSubcomponent
     }