+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/bankauthrepair/BankAuthRepairViewModel.kt
@@ -3,21 +3,24 @@ package com.stripe.android.financialconnections.features.bankauthrepair
 import com.airbnb.mvrx.MavericksViewModel
 import com.airbnb.mvrx.MavericksViewModelFactory
 import com.airbnb.mvrx.ViewModelContext
-import com.stripe.android.financialconnections.features.partnerauth.PartnerAuthState
+import com.stripe.android.financialconnections.features.partnerauth.SharedPartnerAuthState
 import com.stripe.android.financialconnections.model.FinancialConnectionsSessionManifest.Pane
 import com.stripe.android.financialconnections.ui.FinancialConnectionsSheetNativeActivity
 import javax.inject.Inject
 
 internal class BankAuthRepairViewModel @Inject constructor(
-    initialState: PartnerAuthState
-) : MavericksViewModel<PartnerAuthState>(initialState) {
+    initialState: SharedPartnerAuthState
+) : MavericksViewModel<SharedPartnerAuthState>(initialState) {
 
     internal companion object :
-        MavericksViewModelFactory<BankAuthRepairViewModel, PartnerAuthState> {
+        MavericksViewModelFactory<BankAuthRepairViewModel, SharedPartnerAuthState> {
+
+        override fun initialState(viewModelContext: ViewModelContext) =
+            SharedPartnerAuthState(pane = PANE)
 
         override fun create(
             viewModelContext: ViewModelContext,
-            state: PartnerAuthState
+            state: SharedPartnerAuthState
         ): BankAuthRepairViewModel {
             return viewModelContext.activity<FinancialConnectionsSheetNativeActivity>()
                 .viewModel