+++ payments-core/src/test/java/com/stripe/android/view/CardWidgetViewModelTest.kt
@@ -2,12 +2,19 @@ package com.stripe.android.view
 
 import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.PaymentConfiguration
 import com.stripe.android.testing.FeatureFlagTestRule
+import com.stripe.android.utils.FakeCardElementConfigRepository
 import com.stripe.android.utils.FeatureFlags
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import org.junit.Rule
 import org.junit.Test
+
+private val paymentConfig = PaymentConfiguration(
+    publishableKey = "pk_test_123",
+    stripeAccountId = null,
+)
 
 class CardWidgetViewModelTest {
 
@@ -20,29 +27,79 @@ class CardWidgetViewModelTest {
     private val testDispatcher = UnconfinedTestDispatcher()
 
     @Test
-    fun `Emits correct CBC eligibility when the feature is enabled`() = runTest(testDispatcher) {
-        featureFlagTestRule.setEnabled(true)
+    fun `Emits correct CBC eligibility when the feature is enabled and merchant is eligible`() =
+        runTest(testDispatcher) {
+            featureFlagTestRule.setEnabled(true)
 
-        val viewModel = CardWidgetViewModel(
-            cbcEligible = { true },
-            dispatcher = testDispatcher,
-        )
+            val stripeRepository = FakeCardElementConfigRepository()
 
-        viewModel.isCbcEligible.test {
-            assertThat(awaitItem()).isFalse()
-            assertThat(awaitItem()).isTrue()
+            val viewModel = CardWidgetViewModel(
+                paymentConfigProvider = { paymentConfig },
+                stripeRepository = stripeRepository,
+                dispatcher = testDispatcher,
+            )
+
+            viewModel.isCbcEligible.test {
+                assertThat(awaitItem()).isFalse()
+                stripeRepository.enqueueEligible()
+                assertThat(awaitItem()).isTrue()
+            }
         }
-    }
 
     @Test
-    fun `Emits correct CBC eligibility when the feature is disabled`() = runTest(testDispatcher) {
+    fun `Emits correct CBC eligibility when the feature is disabled even if merchant is eligible`() =
+        runTest(testDispatcher) {
+            val stripeRepository = FakeCardElementConfigRepository()
+
+            val viewModel = CardWidgetViewModel(
+                paymentConfigProvider = { paymentConfig },
+                stripeRepository = stripeRepository,
+                dispatcher = testDispatcher,
+            )
+
+            viewModel.isCbcEligible.test {
+                assertThat(awaitItem()).isFalse()
+                stripeRepository.enqueueEligible()
+                expectNoEvents()
+            }
+        }
+
+    @Test
+    fun `Emits correct CBC eligibility when the feature is enabled and merchant is not eligible`() =
+        runTest(testDispatcher) {
+            featureFlagTestRule.setEnabled(true)
+
+            val stripeRepository = FakeCardElementConfigRepository()
+
+            val viewModel = CardWidgetViewModel(
+                paymentConfigProvider = { paymentConfig },
+                stripeRepository = stripeRepository,
+                dispatcher = testDispatcher,
+            )
+
+            viewModel.isCbcEligible.test {
+                assertThat(awaitItem()).isFalse()
+                stripeRepository.enqueueNotEligible()
+                expectNoEvents()
+            }
+        }
+
+    @Test
+    fun `Emits correct CBC eligibility when query fails`() = runTest(testDispatcher) {
+        featureFlagTestRule.setEnabled(true)
+
+        val stripeRepository = FakeCardElementConfigRepository()
+
         val viewModel = CardWidgetViewModel(
-            cbcEligible = { true },
+            paymentConfigProvider = { paymentConfig },
+            stripeRepository = stripeRepository,
             dispatcher = testDispatcher,
         )
 
         viewModel.isCbcEligible.test {
             assertThat(awaitItem()).isFalse()
+            stripeRepository.enqueueFailure()
+            expectNoEvents()
         }
     }
 }